Move transport methods from MoveValidator to TransportUtils
