+++ src/games/strategy/triplea/util/TransportUtils.java
@@ -194,4 +194,17 @@ public class TransportUtils {
     return totalLoad;
   }
 
+  public static int getTransportCost(final Collection<Unit> units) {
+    if (units == null) {
+      return 0;
+    }
+    int cost = 0;
+    final Iterator<Unit> iter = units.iterator();
+    while (iter.hasNext()) {
+      final Unit item = iter.next();
+      cost += UnitAttachment.get(item.getType()).getTransportCost();
+    }
+    return cost;
+  }
+
 }
